Move config to top-level
